chore(user): move user and navigation models to domain package
