+++ app/user/models/__init__.py
@@ -1,0 +1,15 @@
+# app/user/models/__init__.py
+# Domain-owned ORM models for users, permissions, and navigation runtime.
+
+from app.user.models.page_registry import PageRegistry
+from app.user.models.page_route_prefix import PageRoutePrefix
+from app.user.models.permission import Permission
+from app.user.models.user import User, user_permissions
+
+__all__ = [
+    "PageRegistry",
+    "PageRoutePrefix",
+    "Permission",
+    "User",
+    "user_permissions",
+]
