+++ app/user/services/permission_service.py
@@ -7,7 +7,7 @@ from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from app.models.permission import Permission
+from app.user.models.permission import Permission
 
 
 class PermissionService:
